Refactor custom settings handling
to use the builder pattern
